requirements.txt
urllib3==2.2.1
pillow==10.3.0
pytube==15.0.0
ffmpeg-python==0.2.0